Moved functions to new file and changed variable names ingredient_add
